[geometry] moved from xci/core to its own header-only library
